Updating the Russian translation for version 1.21.1 of Neoforge.
Hi, I've updated the Russian translation that was in fashion for you. It was incorrectly localized, but I fixed it. I also noticed one strange error in the keyboard assignment settings menu. In the key assignment settings menu, there is a button called (key.blockprints.open_menu). Perhaps this button is responsible for the configuration of the tools, but it does not have a proper name. Due to this error, it cannot be translated into Russian. Please fix this issue.